Track requests with a unique id
Previously we were losing updates to requests when the returned actions
were different than the first one which originated the request.

Now an additional meta property is added to the actions so that they
will carry through and we can track via mapping to the original request.

We could have also injected a custom `dispatch()` to guarantee the
tracking but this change is simpler and may be safe enough as most
cases so far don't use different actions for the HTTP request
responders.